Relocate validators under tests folder
